+++ test/calex/decoding_test.exs
@@ -221,7 +221,7 @@ defmodule Calex.DecodingTest do
     # https://github.com/nextcloud/calendar/issues/3905#issuecomment-1029970769
 
     data =
-      "BEGIN:VCALENDAR\r\nX-APPLE-STRUCTURED-LOCATION;VALUE=URI;X-ADDRESS=3609 Farmington place\\\\nM\r\n yrtle Beach SC 29579\\\\nUnited States;X-APPLE-ABUID=\"Amanda Loehr’s Home\"\r\n ::;X-APPLE-MAPKIT-HANDLE=CAESiwII2TIaEgkND6uJT9tAQBF6MM6Ey7xTwCJzCg1Vbml\r\n 0ZWQgU3RhdGVzEgJVUxoOU291dGggQ2Fyb2xpbmEiAlNDKgxIb3JyeSBDb3VudHkyDE15cnR\r\n sZSBCZWFjaDoFMjk1NzlSDUZhcm1pbmd0b24gUGxaBDM2MDliEjM2MDkgRmFybWluZ3RvbiB\r\n QbCoSMzYwOSBGYXJtaW5ndG9uIFBsMhIzNjA5IEZhcm1pbmd0b24gUGwyF015cnRsZSBCZWF\r\n jaCwgU0MgIDI5NTc5Mg1Vbml0ZWQgU3RhdGVzODlAAFABWicKJRISCQ0Pq4lP20BAEXowzoT\r\n LvFPAGNkyILKqy+yG3YCw7wGQAwE=;X-APPLE-RADIUS=70.58730101326454;X-APPLE-R\r\n EFERENCEFRAME=1;X-TITLE=3609 Farmington place\nMyrtle Beach SC 29579\nUnited States:geo:33.713365,-78.949922\r\nEND:VEVENT\r\nEND:VCALENDAR\r\n"
+      "BEGIN:VCALENDAR\r\nX-APPLE-STRUCTURED-LOCATION;VALUE=URI;X-ADDRESS=9999 Harmington drive\\\\nM\r\n yrtle Beach SC 29579\\\\nUnited States;X-APPLE-ABUID=\"Aderam Boere’s Home\"\r\n ::;X-APPLE-MAPKIT-HANDLE=CAESiwII2TIaEgkND6uJT9tAQBF6MM6Ey7xTwCJzCg1Vbml\r\n 0ZWQgU3RhdGVzEgJVUxoOU291dGggQ2Fyb2xpbmEiAlNDKgxIb3JyeSBDb3VudHkyDE15cnR\r\n sZSBCZWFjaDoFMjk1NzlSDUZhcm1pbmd0b24gUGxaBDM2MDliEjM2MDkgRmFybWluZ3RvbiB\r\n QbCoSMzYwOSBGYXJtaW5ndG9uIFBsMhIzNjA5IEZhcm1pbmd0b24gUGwyF015cnRsZSBCZWF\r\n jaCwgU0MgIDI5NTc5Mg1Vbml0ZWQgU3RhdGVzODlAAFABWicKJRISCQ0Pq4lP20BAEXowzoT\r\n LvFPAGNkyILKqy+yG3YCw7wGQAwE=;X-APPLE-RADIUS=70.58730101326454;X-APPLE-R\r\n EFERENCEFRAME=1;X-TITLE=9999 Harmington drive\nMyrtle Beach SC 29579\nUnited States:geo:99.999999,-99.999999\r\nEND:VEVENT\r\nEND:VCALENDAR\r\n"
 
     exception =
       try do
@@ -230,7 +230,8 @@ defmodule Calex.DecodingTest do
         e in [Calex.DecodeError] -> e
       end
 
-    assert exception.message == "property has no value: \"Myrtle Beach SC 29579\""
+    assert exception.message ==
+             "property key invalid: \"Myrtle Beach SC 29579\" (reason: invalid_characters)"
   end
 
   test "fails on line with no property key" do
@@ -261,6 +262,29 @@ defmodule Calex.DecodingTest do
     assert_raise(Calex.DecodeError, "property key missing or blank line", fn ->
       Calex.decode!(data)
     end)
+  end
+
+  test "decodes empty values" do
+    data =
+      crlf("""
+      BEGIN:VCALENDAR
+      BEGIN:VEVENT
+      LOCATION:
+      END:VEVENT
+      END:VCALENDAR
+      """)
+
+    assert Calex.decode!(data) == [
+             vcalendar: [
+               [
+                 vevent: [
+                   [
+                     location: {"", []}
+                   ]
+                 ]
+               ]
+             ]
+           ]
   end
 
   test "decodes negative GMT offset dates" do